requirements.txt
browser-use>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0
selenium>=4.0.0
asyncio